requirements.txt
python-multipart
numpy==1.26.4
tensorflow==2.17.0
Pillow
fastapi 
uvicorn

